test defs: more granular overriding of the make program
Before this change, the only way the user could override the make
program used in the automake test cases was to override the $MAKE
variable in the environment.  This had the annoying side effect of
requiring that, whenever a non-default make program was to be used
in the test cases, that same make program had to be used to drive
the execution of the automake testsuite; otherwise, the recursive
make invocations could pick up $MAKE from the environment, and use
that instead of re-executing with the correct make.

So, for example, if one wanted to try how Solaris /usr/ccs/bin/make
behaved in the automake test cases, he couldn't run the testsuite
in parallel mode, because that make lacks support for concurrent
execution of recipes; on fast machines, this easily meant a 4x or
higher slow-down.

Once the problem is clear, the solution is pretty simple: allow
the use of another variable, besides $MAKE, to override the make
program to be used in the test cases.

See also commit `v1.11-1318-g3ceeef4', that introduced a more
general version of this change to the master branch.

* tests/defs.in: Allow the make implementation to be used by the
test cases to be overridden by the `$AM_TESTSUITE_MAKE' variable,
in preference to the `$MAKE' variable.