news: fix botched merges, rebase on 1.11.2a NEWS version
The NEWS file is unfortunately written in a format that doesn't
lend itself very well to automatic merges; the recent merges of
maint into master have left it full of blunders and in a somewhat
inconsistent state.

Another orthogonal issue is that the current version of NEWS
still reports all the changes since the last 1.11 version, rather
than since the latest maintenance version 1.11.2, or, even better,
since the planned next maintenance version 1.11.3.

* NEWS: Fix blunders and errors from botched merges.  Rebase the
file on the NEWS file from branch-1.11 (from which automake 1.11.3
is expected to be released).