tests: work around strangeness in MSYS
MSYS mishandles carriage returns and behaves very strangely for
directories with colon in them.  It seems that colon-directories are
somehow mixed up with drive letters.

Fixes automake bug#7849.

* tests/instspc.tap: Skip instead of erroring out when $test_string
is empty for the carriageret case, as that is expected on MSYS.  Also,
for similar reasons, skip instead of erroring out when it is not
possible to cd into the just created directory, and the directory
name contains a colon.  Update copyright years.