scripts: quote 'like this', not `like this'
This change follows up on recent changes to the GNU coding standards.
They now suggest that we should quote 'like this' or "like this" instead
of `like this'; see:
 <http://www.gnu.org/prep/standards/html_node/Quote-Characters.html>.

Gnulib is being changed accordingly, and Gnulib imports some files
directly from Automake master, so change those files to use the
straight-up style.  This affects only commentary and quoting in
diagnostics.