tests: avoid spurious failures with parallel make
Some tests are unprepared to be run with a make command that runs
in parallel mode by default.  This can happen e.g., if the user
explicitly run the tests with something like:

  AM_TESTSUITE_MAKE="gmake -j3" in the

in the environment, but also if the make used in the testsuite is
Sun Distributed Make, and the 'DMAKE_MODE' environment variable is
set to "parallel".

Fix some instances of this issue.

* tests/tap-doc.test (Makefile.am): Declare explicit dependencies
among the test cases, to ensure they are run in the correct order.
* tests/tap-doc2.test: Set the 'DMAKE_MODE' environment variable
to "serial", to prevent Sun dmake from trying to run in parallel.
Remove other now-superfluous workarounds for dmake.