tap: some makes in parallel mode unconditionally serialize output
Some make implementations (among them, FreeBSD make, NetBSD make,
and Solaris Distributed Make), when run in parallel mode, serialize
the output from their targets' recipes unconditionally.  This has
the nasty side effect that the user won't see the partial results
of a TAP test until the test has terminated.  This is not something
our TAP driver script can work around unfortunately; in fact, the
driver is sending out its output progressively and "in sync" with
test execution -- it is make that is stowing such output away
instead of presenting it to the user as soon as it gets it.

So we content ourself with working around the issue in our
testsuite, to at least avoid failures we could do nothing to
prevent.

* tests/tap-realtime.test: Skip this test if the make program used
is a non-GNU make running in parallel mode.  And if Sun dmake is
being used, try to force it to run in serial mode, by exporting the
'DMAKE_MODE' environment variable to "serial".