dist tests: missing 'compress' program was causing spurious failures
Fixes automake bug#10575.

The compress(1) tool is becoming anachronistic, and thus is not
installed by default on modern distros (e.g., Fedora 16).  We
can't thus unconditionally assume it exists on every reasonable
portability target.

* tests/dist-formats.tap ($missing_compressors): When defining
this, don't assume anymore that compress(1) is unconditionally
available.
(Parallel compression): Skip this sub-test if compress(1) is
unavailable.
Since we are at it, fix a couple of unrelated buglets: a typo
in a test name (s/distcheck/ark-exists/), and some attempts to
remove directories with "rm -f".