[ng] deptrack: remove obsolescent comments
* lib/am/depcomp2.am: Remove some comments describing the why and
the how of the verbosity and organization of the recipes.  Since we
are heavily refactoring and re-organizing the code, such comments
are doomed to get out-of-sync with the new layout and/or behaviour
of the code.  If the need arises, we will re-introduce an updated
version of the comments in a future changeset.
And since we are at it, improve another unrelated comment.