cleanup: remove unused .am file
* lib/am/comp-vars.am: Remove.  This file has stopped being useful
since commit 'Release-1-4d-206-g3334f9a' of 09-04-2001, and should
have been removed back then.  But apparently only its contents were
removed, leaving the file empty but still present in the repository.
Or maybe this situation is  the consequence of a minor blunder in
the conversion from CVS to Git; either way, this is not a big deal.