[ng] yacc, lex: fix subdir VPATH builds
Starting with recent changes, the automake-generated configure
code does not create the '.Po' files for dependency tracking
anymore.  This is good and proper, but unfortunately, when the
'subdir-objects' option is in use *and* a VPATH build is being
performed, it breaks some rules that *implicitly* relied on that
shell code having already created the subdirectory where the
generated files were to be placed.  This happened in particular
with Yacc and Lex rules.  So fix them to instead explicitly
create the required directories, if not already present.

Bug revealed by failures of 'lex5.test' and 'yacc8.test' with
GNU make 3.82 (on all of Debian, Fedora and Solaris).

* lib/am/yacc.am, lib/am/lex.am: Update.
* tests/Makefile.am (XFAIL_TESTS): Remove 'lex-subobj-nodep.test',
as the test now passes.
* tests/yacc-subobj-nodep.test: New test.
* tests/list-of-tests.mk: Add it.