tests: avoid ksh bug for wildcards in EXTRA_DIST
Some shells, like NetBSD 5.1 /bin/ksh, have a bug that causes
wildcards resulting from an unquoted parameter expansion not to
be expanded as expected in some cases:

    $ touch a b c d
    $ /bin/sh -c 'var="[ab]" && echo $var' # As expected.
    a b
    $ /bin/ksh -c 'var="[ab]" && echo $var' # Oops.
    [ab]
    $ /bin/ksh -c 'var="?" && echo $var' # But this as expected.
    a b c d

This was causing a failure in our testsuite.  Work around that.

See also:
<http://lists.gnu.org/archive/html/automake-patches/2012-02/msg00149.html>

* tests/extra11.test: Work around the described issue by using the
$(wildcard) builtin in the EXTRA_DIST definition.  This is not a
problem, since this test was assuming GNU make anyway.  Since we are
at it, enhance coverage a little by also running "make distcheck".
* tests/extra10.test: Similarly, enhance coverage by also running
"make distcheck", and fix a minor bug in the process.  But continue
using a "bare" wildcard expression in EXTRA_DIST, without involving
the $(wildcard) builtin, to ensure coverage of similar usages "in
the wild".