cscope: is only expected to work with GNU make in VPATH setup
The 'cscope' functionality does not properly handle VPATH rewrites;
so we explicitly document that, for now, it is only ensured to work
with GNU make when doing a VPATH build, and we adjust testsuite
requirements accordingly.

Issue revealed by a failure of 'cscope.test' with Solaris 10 CCS
make.

* doc/automake.texi (Tags): Document the limitations of the cscope
support.  Couple of fixlets to formatting since we are at it.
* tests/cscope.test: Converted to TAP, and thus moved ...
* tests/cscope.tap: ... here.  Only run tests with VPATH builds when
using GNU make.  Adjust so that no C, C++, Fortran or Java compiler,
nor any Emacs program, will be required to run this test.
* tests/list-of-tests.mk: Adjust.