[ng] maint: use name "Automake-NG" in more places, bump version
It's time to start renaming this project for good, if we want to
avoid nasty confusions with mainstream Automake.  But renaming also
the 'automake' script to 'automake-ng' could cause gratuitous
incompatibilities and unexpected errors, so we prefer to instead
bump the version from '1.11a' to '2.0a'.

The renaming "Automake" -> "Automake-NG" is mostly for README files
and some comments.  Doing that for the documentation would be
trickier, and might cause extra spurious conflicts when we merge
'master' into 'ng/master', so we leave that move for a later time.

* AUTHORS, README, contrib/README, tests/README, doc/amhello/README:
Update comments, and, where necessary, the mailing list and bug
reporting address.
* doc/Makefile.am, GNUmakefile, syntax-checks.mk, m4/init.m4,
m4/strip.m4, m4/substnot.m4: Update a few comments.
* configure.ac (AC_INIT): Update package name and bug reporting
address.  Reset Package version to "2.0a".
(AM_INIT_AUTOMAKE): Remove minimal required version; we expect
to be bootstrapped anyway.
Minor updates to comments and messages.
* lib/ar-lib: Update the "this file is maintained in Automake" line
to read "Automake-NG" instead.  Also, update the address where to
send patches to (to <automake-ng@gnu.org>), but, at least for the
moment, *not* the bug reporting address.
* lib/compile: Likewise.
* lib/elisp-comp: Likewise.
* lib/mdate-sh: Likewise.
* lib/mkinstalldirs: Likewise.
* lib/py-compile: Likewise.
* lib/tap-driver.pl: Likewise.
* lib/tap-driver.sh: Likewise.
* lib/test-driver: Likewise.