tests: relax a test to cater to NetBSD make
* tests/parallel-tests-empty-testlogs.test: When $(TESTS) is empty,
NetBSD 5.1 make ends up defining $(TESTS_LOGS) to ".log" rather than
to the empty string (as would be expected).  This behaviour was
causing a check in this test case to fail.  But that check was in
fact overly strict, since our recipes are smart enough to work
around the botched-up substitution.  Also, the failure wasn't a
regression from Automake 1.11, since that version too would have
triggered the same error with NetBSD 5.1 make.  So we just relax
the test a little to avoid extra noise in the testsuite.